KissBot/requirements.txt
# KissBot V1 - Requirements officiels TwitchIO 2.x
# Version stable et testée qui FONCTIONNE ✅

# Bot Twitch - Version 2.x STABLE (3.x est buggé pour IRC)
twitchio==2.7.0

# HTTP et APIs
aiohttp==3.13.1
httpx==0.28.1

# Configuration
PyYAML==6.0.3

# LLM et APIs (optionnel)
# requests>=2.31.0  # Pour APIs externes si nécessaire

# Tests unitaires (!trad et autres)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Développement (optionnel)
# black>=23.0.0

# INSTALLATION:
# pip install -r requirements.txt

# INSTALLATION AVEC TESTS:
# pip install -r requirements.txt
# pytest test_translation.py -v

# Note: TwitchIO 3.x a des bugs IRC, utiliser 2.x uniquement!